removed 2 warnings
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@591 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/plugins/loop_sampler/WaveLoop.cpp
@@ -95,8 +95,8 @@ void WaveLoop::SetLoopCursors()
 {
   double inc = (double)GetClientSize().x / (double)EndWavePos;
 
-  LoopStartCursor->SetPosition(wxPoint(inc * LoopInfo->Start, 0));
-  LoopEndCursor->SetPosition(wxPoint(inc * LoopInfo->End, 0));
+  LoopStartCursor->SetPosition(wxPoint((int)(inc * LoopInfo->Start), 0));
+  LoopEndCursor->SetPosition(wxPoint((int)(inc * LoopInfo->End), 0));
 }
 
 void WaveLoop::OnSliceBtnSelected(wxCommandEvent &event)
